backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
djlint==1.36.4
EditorConfig==0.17.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.0
MarkupSafe==3.0.2
pathspec==0.12.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
